+++ build.rs
@@ -1,10 +1,10 @@
-extern crate pkg_config;
+use std::collections::{HashMap, HashSet};
 use std::ffi::OsString;
 use std::process::Command;
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use std::env;
 
-#[cfg(target_os = "windows")]
+#[cfg(not(target_os = "macos"))]
 use std::fs;
 
 macro_rules! ci_stderr_log {
@@ -24,71 +24,6 @@ fn rbconfig(key: &str) -> String {
         .unwrap_or_else(|e| panic!("ruby not found: {}", e));
 
     String::from_utf8(config.stdout).expect("RbConfig value not UTF-8!")
-}
-
-fn ruby_name() -> String {
-    env::var_os("LIBRUBY_NAME")
-        .map(|var| var.to_string_lossy().to_string())
-        .unwrap_or(String::from("ruby"))
-}
-
-fn set_env_pkg_config() {
-    if env::var_os("PKG_CONFIG_PATH").is_none() {
-        let key = "PKG_CONFIG_PATH";
-        let value = Path::new(&rbconfig("libdir")).join("pkgconfig");
-        std::env::set_var(key, &value);
-        ci_stderr_log!("Set PKG_CONFIG_PATH to {:?}", value);
-    }
-}
-
-fn trim_teeny(version: &str) -> &str {
-    version.rsplitn(2, '.').collect::<Vec<&str>>().last().unwrap()
-}
-
-fn ruby_version() -> String {
-    rbconfig("RUBY_PROGRAM_VERSION")
-}
-
-fn transform_lib_args(rbconfig_key: &str, replacement: &str) -> String {
-    rbconfig(rbconfig_key).replace("-l", replacement)
-}
-
-#[cfg(not(target_os = "windows"))]
-fn static_ruby_file_name() -> String {
-    rbconfig("LIBRUBY_A")
-}
-
-#[cfg(target_os = "windows")]
-fn static_ruby_file_name() -> String {
-    let libruby_so = rbconfig("LIBRUBY_SO");
-    let ruby_dll = Path::new(&libruby_so);
-    let name = ruby_dll.file_stem().unwrap();
-    format!("{}.lib", name.to_string_lossy())
-}
-
-#[cfg(not(target_os = "windows"))]
-fn static_ruby_location() -> String {
-    let location = env::var_os("RUBY_STATIC_PATH")
-        .map(|s| s.to_string_lossy().to_string())
-        .unwrap_or(rbconfig("libdir"));
-
-    if !Path::new(&location).join(static_ruby_file_name()).exists() {
-        panic!("{} was not found in path {}, but static build was chosen.\n\
-               Please use environment variable RUBY_STATIC_PATH to define where {} is located.",
-               static_ruby_file_name(), location, static_ruby_file_name());
-    }
-
-    location
-}
-
-#[cfg(target_os = "windows")]
-fn static_ruby_location() -> String {
-    let location: Option<String> = env::var_os("RUBY_STATIC_PATH").map(|s|s.to_string_lossy().to_string());
-    let location: String = location.unwrap_or(
-      Path::new("target").join(env::var_os("PROFILE").unwrap()).join("deps").to_string_lossy().to_string()
-    );
-
-    location
 }
 
 #[cfg(not(target_os = "macos"))]
@@ -138,10 +73,9 @@ fn windows_static_ruby_dep() {
 }
 
 fn use_static() {
-    println!("cargo:rustc-link-search={}", static_ruby_location());
-    let static_name = static_ruby_file_name();
-    let static_name = Path::new(&static_name).file_stem().unwrap().to_string_lossy();
-    println!("cargo:rustc-link-lib={}", static_name.trim_start_matches("lib"));
+    if let Some(location) = env::var_os("RUBY_STATIC_PATH").map(|s|s.to_string_lossy().to_string()) {
+        println!("cargo:rustc-link-search={}", location);
+    }
 
     // If Windows
     windows_static_ruby_dep();
@@ -149,17 +83,15 @@ fn use_static() {
     // If Mac OS
     macos_static_ruby_dep();
 
-    // Ruby gives back the libs in the form: `-lpthread -lgmp`
-    // Cargo wants them as: `-l pthread -l gmp`
     // **Flags must be last in order for linking!**
-    println!("cargo:rustc-flags={}", transform_lib_args("LIBS", "-l "));
+    static_linker_args();
 
     ci_stderr_log!("Using static linker flags");
 }
 
 fn use_dylib() {
     println!("cargo:rustc-link-search={}", rbconfig("libdir"));
-    println!("cargo:rustc-link-lib=dylib={}", rbconfig("RUBY_SO_NAME"));
+    dynamic_linker_args();
     ci_stderr_log!("Using dynamic linker flags");
 }
 
@@ -231,6 +163,183 @@ fn windows_support() {
 #[cfg(not(target_os = "windows"))]
 fn windows_support() {}
 
+#[cfg(target_os = "linux")]
+use std::os::unix::fs::symlink;
+
+#[cfg(target_os = "linux")]
+fn ruby_lib_link_name() -> String {
+    // Rust with linker search paths doesn't seem to use those paths
+    // but rather resorts to the systems Ruby.  So we symlink into
+    // our own deps directory for it to work.
+    let so_file = format!("libruby.so.{}.{}", rbconfig("MAJOR"), rbconfig("MINOR"));
+    let destination = format!("target/{}/deps", env::var("PROFILE").unwrap());
+    let _ = fs::create_dir_all(&destination).map_err(|_|()).expect("create_dir_all fail");
+    let source = format!("{}/{}", rbconfig("libdir"), so_file);
+    let target = format!("{}/{}", destination, so_file);
+
+    if !Path::new(&target).exists() {
+        let _ = symlink(source, target).expect("symlink fail");
+    }
+
+    rbconfig("RUBY_SO_NAME")
+}
+
+#[cfg(target_os = "macos")]
+fn ruby_lib_link_name() -> String {
+    format!(
+      "{}.{}.{}",
+      rbconfig("RUBY_BASE_NAME"),
+      rbconfig("MAJOR"),
+      rbconfig("MINOR")
+    )
+}
+
+#[cfg(target_os = "windows")]
+fn ruby_lib_link_name() -> String {
+    rbconfig("RUBY_SO_NAME")
+}
+
+fn dynamic_linker_args() {
+    let mut library = Library::new();
+    library.parse_libs_cflags(rbconfig("LIBRUBYARG_SHARED").as_bytes(), false);
+    println!("cargo:rustc-link-lib=dylib={}", ruby_lib_link_name());
+    library.parse_libs_cflags(rbconfig("LIBS").as_bytes(), false);
+}
+
+fn static_linker_args() {
+    let mut library = Library::new();
+    library.parse_libs_cflags(rbconfig("LIBRUBYARG_SHARED").as_bytes(), true);
+    library.parse_libs_cflags(format!("-l{}-static", rbconfig("RUBY_SO_NAME")).as_bytes(), true);
+    library.parse_libs_cflags(rbconfig("MAINLIBS").as_bytes(), false);
+}
+
+#[derive(Debug)]
+pub struct Library {
+    pub libs: Vec<String>,
+    pub link_paths: Vec<PathBuf>,
+    pub frameworks: Vec<String>,
+    pub framework_paths: Vec<PathBuf>,
+    pub include_paths: Vec<PathBuf>,
+    pub defines: HashMap<String, Option<String>>,
+    pub version: String,
+    _priv: (),
+}
+
+impl Library {
+    fn new() -> Library {
+        Library {
+            libs: Vec::new(),
+            link_paths: Vec::new(),
+            include_paths: Vec::new(),
+            frameworks: Vec::new(),
+            framework_paths: Vec::new(),
+            defines: HashMap::new(),
+            version: String::new(),
+            _priv: (),
+        }
+    }
+
+    fn parse_libs_cflags(&mut self, output: &[u8], statik: bool) {
+        let mut is_msvc = false;
+        if let Ok(target) = env::var("TARGET") {
+            if target.contains("msvc") {
+                is_msvc = true;
+            }
+        }
+
+        let words = split_flags(output);
+        let parts = words.iter()
+                          .filter(|l| l.len() > 2)
+                          .map(|arg| (&arg[0..2], &arg[2..]))
+                          .collect::<HashSet<_>>();
+
+        let mut dirs = Vec::new();
+        for &(flag, val) in &parts {
+            match flag {
+                "-L" => {
+                    let meta = format!("rustc-link-search=native={}", val);
+                    println!("cargo:{}", &meta);
+                    dirs.push(PathBuf::from(val));
+                    self.link_paths.push(PathBuf::from(val));
+                }
+                "-F" => {
+                    let meta = format!("rustc-link-search=framework={}", val);
+                    println!("cargo:{}", &meta);
+                    self.framework_paths.push(PathBuf::from(val));
+                }
+                "-I" => {
+                    self.include_paths.push(PathBuf::from(val));
+                }
+                "-l" => {
+                    // These are provided by the CRT with MSVC
+                    if is_msvc && ["m", "c", "pthread"].contains(&val) {
+                        continue;
+                    }
+
+                    if is_static() && statik {
+                        let meta = format!("rustc-link-lib=static={}", val);
+                        println!("cargo:{}", &meta);
+                    } else {
+                        let meta = format!("rustc-link-lib={}", val);
+                        println!("cargo:{}", &meta);
+                    }
+
+                    self.libs.push(val.to_string());
+                }
+                "-D" => {
+                    let mut iter = val.split("=");
+                    self.defines.insert(iter.next().unwrap().to_owned(), iter.next().map(|s| s.to_owned()));
+                }
+                _ => {}
+            }
+        }
+
+        let mut iter = words.iter()
+                            .flat_map(|arg| if arg.starts_with("-Wl,") {
+                                 arg[4..].split(',').collect()
+                             } else {
+                                 vec![arg.as_ref()]
+                             });
+        while let Some(part) = iter.next() {
+            if part != "-framework" {
+                continue
+            }
+            if let Some(lib) = iter.next() {
+                let meta = format!("rustc-link-lib=framework={}", lib);
+                println!("cargo:{}", &meta);
+                self.frameworks.push(lib.to_string());
+            }
+        }
+    }
+}
+
+fn split_flags(output: &[u8]) -> Vec<String> {
+    let mut word = Vec::new();
+    let mut words = Vec::new();
+
+    for &b in output {
+        match b {
+            b' ' => {
+                if !word.is_empty() {
+                    words.push(String::from_utf8(word).unwrap());
+                    word = Vec::new();
+                }
+            }
+            _ => word.push(b),
+        }
+    }
+
+    if !word.is_empty() {
+        words.push(String::from_utf8(word).unwrap());
+    }
+
+    words
+}
+
+fn is_static() -> bool {
+    env::var_os("RUBY_STATIC").is_some()
+}
+
 fn main() {
     // Ruby programs calling Rust doesn't need cc linking
     if let None = std::env::var_os("NO_LINK_RUTIE") {
@@ -238,24 +347,7 @@ fn main() {
         // If windows OS do windows stuff
         windows_support();
 
-        if env::var_os("RUTIE_NO_PKG_CONFIG").is_none() && env::var_os("RUBY_STATIC").is_none() {
-            // Ruby often includes pkgconfig under their lib dir
-            set_env_pkg_config();
-
-            let ruby_version = ruby_version();
-            let version = trim_teeny(&ruby_version);
-
-            // To disable the use of pkg-config set the environment variable `RUTIE_NO_PKG_CONFIG`
-            match pkg_config::Config::new().atleast_version(version).probe(&ruby_name()) {
-                Ok(_) => {
-                    ci_stderr_log!("pkg-config is being used");
-                    return;
-                },
-                Err(err) => ci_stderr_log!("{:?}", err),
-            }
-        }
-
-        if env::var_os("RUBY_STATIC").is_some() {
+        if is_static() {
             ci_stderr_log!("RUBY_STATIC is set");
             use_static()
         } else {