+++ CHANGELOG.md
@@ -6,7 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 for the public APIs. `rubysys`, even though shared publicly, is considered a private
 API and may have breaking changes during a teeny version change.
 
+
 ## [Unreleased]
+### Added
+- `GC::adjust_memory_usage`, thanks to @Antti
+
+### Fixed
+- `GC::mark` documentation notes.
+
+## [0.5.3] - 2019-01-10
 ### Added
 - `util::is_proc` & `util::is_method`
 - `rb_enc_compatible` useful for internal string encoding compatibilty checks from
@@ -27,6 +35,10 @@ API and may have breaking changes during a teeny version change.
 - Removed Ruby 2.3 support & added 2.6
 - `VM::raise_ex` now accepts `Into<AnyException>` rather than just `AnyException`
 - Refactor internal encoding types
+- Refactor `build.rs` script to use Ruby provided cflags
+
+### Removed
+- pkg-config-rs removed from Rutie and from the build process
 
 ## [0.5.2] - 2018-12-18
 ### Added